# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "progettosauri"
  description: Progettosauri SE2 project - API documentation
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: progettosauri-se2-project.herokuapp.com
basePath: /api/v1
consumes:
- application/json
produces:
- application/json
...
paths:
...
  /tasks:
    # Visualize available tasks
    get:
      summary: "View tasks created by and shared with the logged user"
      tags:
        - "Tasks"
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            type: object
            properties:
              created:
                type: array
                items:
                  $ref: "#/definitions/Task"
              shared:
                type: array
                items:
                  $ref: "#/definitions/Task"
        400:
          description: Bad parameter

    # Create tasks
    post:
      summary: "Create a task"
      tags:
        - "Tasks"
      consumes:
        - application/json
      parameters:
        - in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/TaskPost"
      responses:
        201:
          description: Success, task has been created
        400:
          description: Bad parameters
  /tasks/{task_id}:
    # Visualize a task
    get:
      summary: "View details of a task"
      tags:
        - "Tasks"
      parameters:
        - in: path
          name: task_id
          required: true
          type: integer
          minimum: 1
          description: Id of the task
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Task"
        403:
          description: Permission denied
        404:
          description: Task does not exist

    # Update a task
    put:
      summary: "Update a task"
      tags:
        - "Tasks"
      consumes:
        - application/json
      parameters:
        - in: path
          name: task_id
          required: true
          type: integer
          minimum: 1
          description: Id of the task
        - in: body
          name: task
          required: true
          schema:
            $ref: "#/definitions/TaskPut"
      produces:
        - application/json
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/Task"
        403:
          description: Permission denied
        404:
          description: Task does not exist

    # Delete a task
    delete:
      summary: "Delete a task"
      tags:
        - "Tasks"
      parameters:
        - in: path
          name: task_id
          required: true
          type: integer
          minimum: 1
          description: Id of the task
      responses:
        200:
          description: Success
        400:
          description: Bad parameters
        403:
          description: Permission denied
        404:
          description: Task does not exist

######################################################################   Classes
...
definitions:
...
  Task:
    type: object
    properties:
      id:
        type: integer
        description: "Identifier of the task"
      text:
        type: string
        description: "The question of the task"
      answers:
        type: array
        description: "List of possible answers. If it contains less than two answers the task will be an open answer task, otherwise a multiple chioce task"
        items:
          type: object
          description: "A single answer, it can be an open answer or a multiple choice answer"
          properties:
            id:
              type: integer
              description: "Identifier of the answer"
            text:
              type: string
              description: "The text of the answer"
            correct:
              type: boolean
              description: "Correctness of the answer"
  TaskPost:
    type: object
    properties:
      text:
        type: string
        description: "The question of the task"
      answers:
        type: array
        description: "List of possible answers. If it contains less than two answers the task will be an open answer task, otherwise a multiple chioce task"
        items:
          type: object
          description: "A single answer, it can be an open answer or a multiple choice answer"
          properties:
            text:
              type: string
              description: "The text of the answer"
            correct:
              type: boolean
              description: "Correctness of the answer"
          required:
            - text
            - correct
    required:
      - text
  TaskPut:
    type: object
    properties:
      text:
        type: string
        description: "The question of the task"
      answers:
        type: array
        description: "List of possible answers. If it contains less than two answers the task will be an open answer task, otherwise a multiple chioce task"
        items:
          type: object
          description: "A single answer, it can be an open answer or a multiple choice answer"
          properties:
            text:
              type: string
              description: "The text of the answer"
            correct:
              type: boolean
              description: "Correctness of the answer"
...